third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/005-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: Cannot read property 'contentWindow' of null
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: Cannot read property 'contentWindow' of null
Harness: the test ran to completion.
